script/pyproject.toml
[project]
name = "script"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "jwcrypto>=1.5.6",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
]
